// types/static-types/static-types.ts
export type StaffType =
  | "Doctor"
  | "Consultant"
  | "Support"
  | "Nurse"
  | "Technician"
  | "Therapist"
  | "Pharmacist"
  | "Administrator"
  | "Receptionist"
  | "Lab Assistant"
  | "Radiologist"
  | "Surgeon"
  | "Paramedic"
  | "Intern"
  | "Manager"
  | "HR"
  | "Finance"
  | "IT Support"
  | "Operations"
  | "Quality Assurance"
  | "Medical Officer"
  | "Care Coordinator"
  | "Clinical Assistant"
  | "Supervisor"
  | "Director"
  | "Other";

export type AppointmentStatus =
  | "Scheduled"
  | "Completed"
  | "Cancelled"
  | "No-Show"
  | "Queued";